refactor: :truck: move navLinks && update imports
