+++ components/Layout/Layout.tsx
@@ -2,7 +2,7 @@ import { FC, ReactNode } from 'react'
 import { Box } from '@mantine/core'
 import CustomHeader from '../CustomHeader/CustomHeader'
 import Footer from '../Footer/Footer'
-import { navLinks } from '../../routing/navLinks'
+import { navLinks } from '../CustomHeader/navLinks'
 
 interface ILayoutProps {
   children: ReactNode
